feat: add-ts-declaration-files (#6)
* feat: add declaration file

* chore: update repo url
