+++ .changeset/strong-wombats-cross.md
@@ -1,0 +1,5 @@
+---
+'rollup-plugin-sourcemaps2': minor
+---
+
+chore: add declaration files
