+++ package.json
@@ -16,7 +16,7 @@
   "sideEffects": false,
   "repository": {
     "type": "git",
-    "url": "https://github.com/2wce/rollup-plugin-sourcemaps"
+    "url": "https://github.com/2wce/rollup-plugin-sourcemaps2"
   },
   "engines": {
     "node": ">=18.0.0"
@@ -28,7 +28,7 @@
     "prebuild": "npm run lint && npm run format && npm run typecheck",
     "lint": "eslint .",
     "format": "prettier --write .",
-    "typecheck": "tsc --noEmit",
+    "typecheck": "tsc",
     "prepublishOnly": "npm run build",
     "preversion": "npm run test"
   },
